modules working, functional style all over the place!!
