+++ Model/src/main/scala/model/levelComponent/LevelInterface.scala
